+++ SportsHub/SportsHub.xcdatamodeld/SportsHub.xcdatamodel/contents
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="1" systemVersion="11A491" minimumToolsVersion="Automatic" sourceLanguage="Swift" usedWithCloudKit="false" userDefinedModelVersionIdentifier="">
-    <elements/>
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="22757" systemVersion="23E214" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+    <entity name="LeagueEntitie" representedClassName="LeagueEntitie" syncable="YES" codeGenerationType="class">
+        <attribute name="leagueKey" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
+        <attribute name="leagueLogo" optional="YES" attributeType="Binary"/>
+        <attribute name="leagueName" optional="YES" attributeType="String"/>
+    </entity>
 </model>
